+++ ocrd_segment/repair.py
@@ -41,7 +41,7 @@ from ocrd_validators.page_validator import (
     PageValidator
 )
 from .config import OCRD_TOOL
-from .project import join_polygons
+from .project import join_polygons, make_valid
 
 TOOL = 'ocrd-segment-repair'
 
@@ -115,6 +115,14 @@ class RepairSegmentation(Processor):
             pcgts.set_pcGtsId(file_id)
             page = pcgts.get_Page()
 
+            # shrink/expand text regions to the hull of their text lines
+            if sanitize:
+                page_image, page_coords, _ = self.workspace.image_from_page(
+                    page, page_id,
+                    feature_selector='binarized',
+                    feature_filter='clipped')
+                shrink_regions(page_image, page_coords, page, page_id,
+                               padding=self.parameter['sanitize_padding'])
             #
             # validate segmentation (warn of children extending beyond their parents)
             #
@@ -180,14 +188,6 @@ class RepairSegmentation(Processor):
             # delete/merge/split redundant text regions (or its text lines)
             if plausibilize:
                 self.plausibilize_page(page, page_id)
-            # shrink/expand text regions to the hull of their text lines
-            if sanitize:
-                page_image, page_coords, _ = self.workspace.image_from_page(
-                    page, page_id,
-                    feature_selector='binarized',
-                    feature_filter='clipped')
-                shrink_regions(page_image, page_coords, page, page_id,
-                               padding=self.parameter['sanitize_padding'])
 
             self.workspace.add_file(
                 ID=file_id,
@@ -482,7 +482,7 @@ def _plausibilize_segments(segpolys, rogroup, marked_for_deletion, marked_for_me
                      _tag_name(otherseg), otherseg.id)
             otherpoly = make_valid(Polygon(polygon_from_points(otherseg.get_Coords().points)))
             poly = poly.difference(otherpoly)
-            if poly.type == 'MultiPolygon':
+            if poly.geom_type == 'MultiPolygon':
                 poly = join_polygons(poly.geoms)
             if poly.minimum_clearance < 1.0:
                 poly = Polygon(np.round(poly.exterior.coords))
@@ -556,8 +556,8 @@ def shrink_regions(page_image, page_coords, page, page_id, padding=0):
             continue
         # pick contour and convert to absolute:
         region_polygon = join_polygons([make_valid(Polygon(contour[:, 0, ::]))
-                                        for contour in contours
-                                        if len(contour) >= 3], scale=scale)
+                                        for area, contour in zip(areas, contours)
+                                        if len(contour) >= 3 and area > 0], scale=scale)
         if padding:
             region_polygon = region_polygon.buffer(padding)
         region_polygon = coordinates_for_segment(region_polygon.exterior.coords[:-1], page_image, page_coords)
@@ -599,7 +599,7 @@ def simplify(segment, tolerance=0):
 
 def merge_poly(poly1, poly2):
     poly = poly1.union(poly2)
-    if poly.type == 'MultiPolygon':
+    if poly.geom_type == 'MultiPolygon':
         #poly = poly.convex_hull
         poly = join_polygons(poly.geoms)
     if poly.minimum_clearance < 1.0:
@@ -611,10 +611,10 @@ def clip_poly(poly1, poly2):
     poly = poly1.intersection(poly2)
     if poly.is_empty or poly.area == 0.0:
         return None
-    if poly.type == 'GeometryCollection':
+    if poly.geom_type == 'GeometryCollection':
         # heterogeneous result: filter zero-area shapes (LineString, Point)
         poly = unary_union([geom for geom in poly.geoms if geom.area > 0])
-    if poly.type == 'MultiPolygon':
+    if poly.geom_type == 'MultiPolygon':
         # homogeneous result: construct convex hull to connect
         #poly = poly.convex_hull
         poly = join_polygons(poly.geoms)
@@ -718,21 +718,6 @@ def ensure_valid(element):
     if changed:
         points = points_from_polygon(polygon)
         coords.set_points(points)
-
-def make_valid(polygon):
-    """Ensures shapely.geometry.Polygon object is valid by repeated simplification"""
-    for split in range(1, len(polygon.exterior.coords)-1):
-        if polygon.is_valid or polygon.simplify(polygon.area).is_valid:
-            break
-        # simplification may not be possible (at all) due to ordering
-        # in that case, try another starting point
-        polygon = Polygon(polygon.exterior.coords[-split:]+polygon.exterior.coords[:-split])
-    for tolerance in range(1, int(polygon.area)):
-        if polygon.is_valid:
-            break
-        # simplification may require a larger tolerance
-        polygon = polygon.simplify(tolerance)
-    return polygon
 
 def _tag_name(element):
     return element.__class__.__name__[0:-4]